+++ logic/mqtt/MqttThreadManager.py
@@ -17,5 +17,5 @@ class MqttThreadManager:
 
     def _mqtt_worker(self):
         self.mqtt_client.start_loop()
-        time.sleep(5)
+        time.sleep(10)
         self.mqtt_client.stop_loop()
