+++ resources/layout.kv
@@ -84,6 +84,15 @@
             opacity: 1
             disabled: False
 
+        Button:
+            text: '⟳'
+            size_hint: None, None
+            size: dp(60), dp(60)
+            pos_hint: {'center_x': 0.5, 'center_y': 0.5}
+            on_release: root.sync_items()
+            opacity: 1
+            disabled: False
+
 <LoginScreen>:
     name: 'login'
     BoxLayout:
