+++ resources/tasks.json
@@ -9,7 +9,7 @@
         {
             "uuid": "6e6eb9b4-99f7-4d18-b59b-271ccce17e48",
             "client-id": "e4947a81-dbc2-40de-aad8-0dd3eab591f0",
-            "message": "New Task",
+            "message": "New Taskhallo",
             "state": false
         },
         {
@@ -21,7 +21,7 @@
         {
             "uuid": "92b402f2-4502-411e-a9b2-d3415addfc99",
             "client-id": "e4947a81-dbc2-40de-aad8-0dd3eab591f0",
-            "message": "New Task",
+            "message": "New TaskMoin",
             "state": false
         },
         {
@@ -32,6 +32,18 @@
         },
         {
             "uuid": "709958de-0c58-433f-ae3f-209b2a6648a0",
+            "client-id": "e4947a81-dbc2-40de-aad8-0dd3eab591f0",
+            "message": "New Task",
+            "state": false
+        },
+        {
+            "uuid": "0369bcfd-2c78-4d25-a7eb-88f6421cfc4b",
+            "client-id": "e4947a81-dbc2-40de-aad8-0dd3eab591f0",
+            "message": "New Task",
+            "state": false
+        },
+        {
+            "uuid": "1a2096da-0cf2-4672-835e-76da2e209513",
             "client-id": "e4947a81-dbc2-40de-aad8-0dd3eab591f0",
             "message": "New Task",
             "state": false
