+++ ui/components/NavigationBar.py
@@ -1,0 +1,4 @@
+from kivymd.uix.behaviors import FakeRectangularElevationBehavior
+from kivymd.uix.floatlayout import MDFloatLayout
+class NavBar(FakeRectangularElevationBehavior, MDFloatLayout):
+    pass
